backup_local/requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
PyYAML>=6.0
typer>=0.7.0
streamlit>=1.28.0 